Rename tsoclient package to tso.client
