+++ examples/src/main/resources/hbase-omid-client-config.yml
@@ -3,7 +3,7 @@
 # principal: omid_hbase_client
 # keytab: /path/to/hbase/client/keytab
 
-omidClientConfiguration: !!com.yahoo.omid.tsoclient.OmidClientConfiguration
+omidClientConfiguration: !!com.yahoo.omid.tso.client.OmidClientConfiguration
     connectionString: "localhost:54758"
 
 metrics: !!com.yahoo.omid.metrics.CodahaleMetricsProvider [
