+++ transaction-client/src/main/java/com/yahoo/omid/tso/client/NewTSOException.java
@@ -15,14 +15,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.yahoo.omid.tsoclient;
+package com.yahoo.omid.tso.client;
 
 /**
- * Thrown when some incompatibilities between the TSO client & server are
- * found
+ * Thrown when a new TSO has been detected
  */
-public class HandshakeFailedException extends Exception {
+public class NewTSOException extends Exception {
 
-    private static final long serialVersionUID = 8545505066920548834L;
+    private static final long serialVersionUID = -3250655858200759321L;
 
 }
