+++ transaction-client/src/main/resources/omid-client-config.yml
@@ -7,13 +7,13 @@
 # ---------------------------------------------------------------------------------------------------------------------
 
 # Direct connection to host:port
-connectionType: !!com.yahoo.omid.tsoclient.OmidClientConfiguration$ConnType DIRECT
+connectionType: !!com.yahoo.omid.tso.client.OmidClientConfiguration$ConnType DIRECT
 connectionString: "localhost:54758"
 
 # When Omid is working in High Availability mode, two or more replicas of the TSO server are running in primary/backup
 # mode. When a TSO server replica is elected as master, it publishes its address through ZK. In order to configure
 # the Omid client to access the TSO server in HA mode:
-#     1) set 'connectionType' to !!com.yahoo.omid.tsoclient.OmidClientConfiguration$ConnType HA
+#     1) set 'connectionType' to !!com.yahoo.omid.tso.client.OmidClientConfiguration$ConnType HA
 #     2) set 'connectionString' to the ZK cluster connection string where the server is publishing its address
 zkConnectionTimeoutInSecs: 10
 # In HA mode, make sure that the next settings match same settings on the TSO server side
@@ -36,4 +36,4 @@ executorThreads: 3
 
 # Configure whether the TM performs the post-commit actions for a tx (update shadow cells and clean commit table entry)
 # before returning to the control to the client (SYNC) or in parallel (ASYNC)
-postCommitMode: !!com.yahoo.omid.tsoclient.OmidClientConfiguration$PostCommitMode SYNC
+postCommitMode: !!com.yahoo.omid.tso.client.OmidClientConfiguration$PostCommitMode SYNC
