+++ tso-server/src/test/java/com/yahoo/omid/tso/client/TestTSOClientResponseHandling.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.yahoo.omid.tsoclient;
+package com.yahoo.omid.tso.client;
 
 import com.yahoo.omid.tso.ProgrammableTSOServer;
 import com.yahoo.omid.tso.ProgrammableTSOServer.AbortResponse;
